library/spi_engine: Update for SDO extension
* Extend SDO support to 8 (symmetrical with SDI support);
* Update SDI to use asymmetrical FIFO;
* Insert symmetrical FIFO for the SDO;
* Insert SDI lane mask configuration instruction to reg 3'b011;
* Insert SDO lane mask configuration instruction to reg 3'b100;
* Insert offload active interface for interconnect and execution;
* Remove register 8'h3b from spi engine regmap;
* Improvements on the critical paths of the execution module.

Prefetching on offload work iff all lanes are active.

There is a different register for SDI lane mask and SDO
lane mask;
Update the NUM_OF_SDI parameter to NUM_OF_SDIO parameter to reflect
that both SDI and SDO are symmetrical. The control over them is
through the SDI/SDO lane masks.

Inserts a ready signal for when the valid_indices
has finished its inner logic. This avoid the possible
issue where the latency of the command is smaller than
this logic. This could only happen in the FIFO mode.

Introduces logic to keep track of the number of elements in
the SDI FIFO. This is due to the asymmetric FIFO implementation
not including the logic for this, and instead outputting the
level of one of its internal FIFOs.

The modifications performed in this library are not backwards compatible. So
this version is Spi Engine 2.0.

SRC_REG_SLICE_EN enables the register slice inside the
FIFO ASYM, this is necessary for closing time in the AD4052
and AD4630 projects.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>
Signed-off-by: Carlos Souza <carlos.souza@analog.com>