projects/ad4170; projects/ad469x_evb: simplify unneeded logic
ad4170_asdz and ad469x_evb both used some logic at the top level to synchronize
a signal from the ADC to the SPI Engine domain, and then to detect a falling
edge on the synchronized signal, using it as a trigger for the SPI Engine.
Recently, the SPI Engine was updated to use an edge-based trigger, with the
appropriate synchronizer logic added, so this extra logic is no longer needed.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>