library/spi_engine: update spi lane mask instruction
There is a different register for SDI lane mask and SDO
lane mask;
Update the NUM_OF_SDI parameter to NUM_OF_SDIO parameter to reflect
that both SDI and SDO are symmetrical. The control over them is
through the SDI/SDO lane masks.

Inserts a ready signal for when the valid_indices
has finished its inner logic. This avoid the possible
issue where the latency of the command is smaller than
this logic. This could only happen in the FIFO mode.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>