library/spi_engine: fix execution
Fixes the bug related to the sdo lane config instruction, which
was high all the time and not for a single cycle. That was causing
the data_assemble module to be rebuilding the lookup table, and
consequently affecting the output of the offload mode.

Removed lane_index_d because it is not necessary anymore. Both FIFO
and OFFLOAD modes are using the valid_indices lookup table.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>