library: fix wrong reset polarity role for altera IP (#1826)
Quartus detects reset polarity from the signal role, and inserts adapaters
automatically. The IP affected by this commit were previously assigning the
wrong (active high) polarity to active-low resets. This can cause multiple
problems.

The i3c controller and spi engine resets are generated by other IPs in the
framework, and these were also detected as wrong polarity, hence this being
undetected so far.

ad4170_asdz and ad469x_evb both used some logic at the top level to synchronize
a signal from the ADC to the SPI Engine domain, and then to detect a falling
edge on the synchronized signal, using it as a trigger for the SPI Engine.
Recently, the SPI Engine was updated to use an edge-based trigger, with the
appropriate synchronizer logic added, so this extra logic is no longer needed.
This is done in this same commit instead of adjusting the unneeded logic to the
new corrected reset polarity.

Signed-off-by: Laez Barbosa <laez.barbosa@analog.com>