library/util_axis_fifo_asym: fix hdl for quartus
Fifo asym library is only being used by data_offload, which is not
instantiated in quartus projects. However, future spi engine 2.0 will
use it and it will break compilation.

Quartus demands a label for every for loop and this commit fixes it. No
logic update was done in the HDL, so that is not going to change the
current behavior of this library.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>