versal: Add JESD204 support for GTM transceivers
Adds GTM transceiver support for the JESD204 framework.

The GTM's don't have an internal 64b66b gearbox for
JESD204C so an external one that's also being used on
the Agilex projects is used.

For JESD204B we are using the soft_pcs implementation
because of the lack of 8b10b encoders and decoders.

All the projects that are using the versal_xcvr_subsystem
script have been updated.

Signed-off-by: Bogdan Luncan <bogdan.luncan@analog.com>