projects/ad4630_fmc: Refactor ad463x
Separates the 1CH (AD4030) HDL support from the 2CH (AD4630) support.
Previously, whenever 1CH was used it was necessary to throw away half
of the data because the project was always expecting 2 channels for the
ADC.

Updated the XDC files for each supported mode considering how the amount
of channels the ADC. Also updated the tcl script variables, now there is a
NUM_OF_CHANNEL, updated the NUM_OF_SDI to LANES_PER_CHANNEL, and
inserted INTERLEAVE_MODE variable. NO_REORDER became an internal compilation
variable that is enabled according to the combination NUM_OF_CHANNEL,
LANES_PER_CHANNEL and INTERLEAVE_MODE. Setting INTERLEAVE_MODE to 1 with
unsupported configuration throws an error.

Inserts a GPIO to drive CNV pin in parallel with the PWM. The CNV pin is an
"or" function of the GPIO[36] and PWM.

Updated examples in the README FILE.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>