Update: tcl scripts and insert GPIO
Previous tcl scripts allowed the usage of NO_REORDER flag with
unsupported modes. Current tcl script use INTERLEAVE_MODE flag,
that defines when the reorder IP is necessary or not. Setting
INTERLEAVE_MODE to 1 with unsupported configuration throws an
error.

Inserts a GPIO to drive CNV pin in parallel with the PWM. The
CNV pin is an or function of the GPIO[36] and PWM.

Updated examples in the README FILE.

Signed-off-by: Carlos Souza <carlos.souza@analog.com>