.github/scripts/check_readme.sh: Added special project case
- In the previous implementation, the script would throw an error
if the project missed a certain section in the project/README.md file
(e.g. No "Supported parts" section)
- To solve this issue and not introduce a new flag (the project/README.md
files don't include any flags) this script was updated to include a function
for 'special' projects (special = projects which can't respect the README.md
files template). Simply add the 'special' project to the the list, here in the
script, thus the workflows won't fail saying that the README.md files don't
respect the template.
- Also updated the .github/scripts/README.md file to reflect the changes

Signed-off-by: Cristian Mihai Popa <cristianmihai.popa@analog.com>