library: xilinx: scripts: versal_xcvr_subsystem: Clarity updates and support for non-consecutive quads
Removed duplicate code by wrapping code fragments in functions.
Non-consecutive quad mode is needed by AD9084 because its sides
are mapped backwards to the quads (side B goes to the first one,
side A goes to the second one). The Versal Transceiver Subsystem
doesn't allow you to map Quad0 to Quad1 and vice-versa so you
need to instantiate two of them with one quad each.

This method also allows us to build the VCK190 design for
side B only or to build the VCK190/VPK180 designs with
less than 4 lanes.

Signed-off-by: Bogdan Luncan <bogdan.luncan@analog.com>